join byte and master

+++ src/main/java/util/DataContainers/EvolutionResults.java
@@ -8,7 +8,7 @@ import java.util.ArrayList;
 public record EvolutionResults(
         int generation,
         Agent bestAgent,
-        long bestScore,
-        long[] scores
+        int bestScore,
+        int[] scores
 ) {
 }
