+++ src/main/java/util/helpers/EvalFunc.java
@@ -2,5 +2,5 @@ package util.helpers;
 
 @FunctionalInterface
 public interface EvalFunc<T> {
-    long eval(T obj);
+    int eval(T obj);
 }
